Change conky font size to 9pt
Conky font size appears to be in pt, and is effected by desktop scaling,
while cairo is in px and is not effected by desktop scaling. Based on a
typical 96 DPI configuration, use a non-scaled 9pt in conky, such that
the scaled 12px font in cairo will be the same size.